moved elevation profile files to correct folder
