+++ src/modules/elevationProfile/index.js
@@ -1,5 +1,5 @@
 import './i18n';
-import { ElevationProfile } from './components/ElevationProfile';
+import { ElevationProfile } from './components/panel/ElevationProfile';
 
 if (!window.customElements.get(ElevationProfile.tag)) {
 	window.customElements.define(ElevationProfile.tag, ElevationProfile);
